feat(circuit): select parser backend via DEQ_PARSER; share body validation
Wire deqagram behind parser.parse(): the DEQ_PARSER env var selects "lark"
(default) or "deqagram". parse_file / render_and_parse_* and _resolve_file all
route through parse(), so import resolution works under either backend
unchanged. The deqagram branch goes through the model.py shim and maps its
ValueError to SyntaxError to match deq's contract.

Extract the transformer's body-level semantic validators (port ordering,
CONDITIONAL/PROPAGATE-after-OUTPUT, PRESELECT placement, ports-in-REPEAT) into a
new parser-agnostic module deq/circuit/body_validation.py that operates on
model.py objects. The lark transformer now imports these (behavior unchanged:
its passing test set is identical), and the deqagram shim calls them too so both
backends enforce the same rules. The shim also applies the REPEAT count>=1 check.

With this, deq's full test suite passes the exact same 382 tests under
DEQ_PARSER=deqagram as under lark (the remaining failures/collection errors are
pre-existing, from unrelated binar/deq_runtime dependency mismatches in this
environment). This includes the previously diverging cases: port-in-REPEAT,
CONDITIONAL/PROPAGATE ordering, decorator string escapes (fixed in deqagram),
and PROPAGATE-without-FROM (rejected by port ordering).

Copilot-Session: 95863b1b-8294-40a9-a872-33a937c8439f
